refactor: one file for each modal and re-use ApproveRejectModal
